Working Program
(Meant to push this with the original commit) Drop .py executable into a folder containing a season of TV, run it, enter the season number when prompted and this program will rename them into a Plex friendly syntax. Obviously the video types are not complete (.avi anad .mp4) but they can be easily added.